Move UserCreate.tsx to try to solve folder case issue
